requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
langchain-text-splitters
chromadb
streamlit
faiss-cpu
langchain-ollama